requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
sqlalchemy-utils
numpy
pgvector
openai
